Extract SQL helper input normalization into SQLAsset input module
