+++ docs/lib_structure.md
@@ -77,6 +77,7 @@ lib/
     │   ├── compiler.ex
     │   ├── definition.ex
     │   ├── error.ex
+    │   ├── input.ex
     │   ├── materialization.ex
     │   ├── renderer.ex
     │   └── runtime.ex
